pyproject.toml
[tool.poetry]
name = "tageditor"
version = "0.1.0"
description = "Tool for edit audio tags"
authors = ["dimaevan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytaglib = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
